refact: using custom hook
